+++ cypress/support/e2e.js
@@ -1,4 +1,2 @@
-import { enableAutoLogin } from '@dhis2/cypress-commands';
+import '@dhis2/cypress-commands';
 import './commands';
-
-enableAutoLogin();