+++ src/core_modules/capture-core/components/DataEntries/SingleEventRegistrationEntry/DataEntryWrapper/DataEntry/epics/saveNewSingleEventAddAnother.epics.js
@@ -12,9 +12,6 @@ import {
 
 import {
     newRecentlyAddedEvent,
-} from '../../RecentlyAddedEventsList/recentlyAddedEventsList.actions';
-
-import {
     prependListItem,
     removeListItem,
 } from '../../RecentlyAddedEventsList';