+++ src/core_modules/capture-core/components/PossibleDuplicatesDialog/possibleDuplicatesDialog.epics.js
@@ -84,9 +84,10 @@ export const loadSearchGroupDuplicatesForReviewEpic = (
                     ...contextParam,
                 };
                 const attributes = getAttributesFromScopeId(selectedScopeId);
+                const programId = scopeType === scopeTypes.TRACKER_PROGRAM ? selectedScopeId : null;
 
                 const stream$: Stream = from(
-                    getTrackedEntityInstances(queryArgs, attributes, absoluteApiPath, querySingleResource),
+                    getTrackedEntityInstances(queryArgs, attributes, absoluteApiPath, querySingleResource, programId),
                 );
                 return stream$.pipe(
                     map(({ trackedEntityInstanceContainers: searchResults, pagingData }) =>