+++ src/core_modules/capture-core/components/Relationships/Relationships.component.js
@@ -145,6 +145,7 @@ class RelationshipsPlain extends React.Component<Props> {
                         wrapperClassName={classes.tooltip}
                     >
                         <IconButton
+                            data-test="delete-relationship-button"
                             onClick={() => { onRemoveRelationship(relationship.clientId); }}
                             disabled={!canDelete}
                         >